@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core";
 import { Restaurant } from "./restaurant/restaurant.model";
 import { MEATT_API } from "../app.api";
-import { Http } from "@angular/http";
+import { HttpClient, HttpParams } from "@angular/common/http";
 import { Observable } from "rxjs/Observable";
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
@@ -11,31 +11,28 @@ import { MenuItem } from "../restaurant-detail/menu-item/menu-item.model";
 @Injectable()
 export class RestaurantService {
 
-    constructor(private http: Http) {
+    constructor(private http: HttpClient) {
 
     }
 
     restaurants(search?: string): Observable<Restaurant[]> {
-        return this.http.get(`${MEATT_API}/restaurants`, { params: { q: search } })
-            .map(response => response.json())
+        let params: HttpParams = undefined;
+        if (search) {
+            params = new HttpParams().append('q', search);
+        }
+        return this.http.get<Restaurant>(`${MEATT_API}/restaurants`, { params: params })
             .catch(ErrorHandler.handleError);
     }
 
     restaurantById(id: string): Observable<Restaurant> {
-        return this.http.get(`${MEATT_API}/restaurants/${id}`)
-            .map(res => res.json())
-            .catch(ErrorHandler.handleError);
+        return this.http.get<Restaurant>(`${MEATT_API}/restaurants/${id}`)
     }
 
     reviewsOfRestaurant(id: string): Observable<any> {
         return this.http.get(`${MEATT_API}/restaurants/${id}/reviews`)
-            .map(res => res.json())
-            .catch(ErrorHandler.handleError);
     }
 
     menuOfRestaurant(id: string): Observable<MenuItem[]> {
-        return this.http.get(`${MEATT_API}/restaurants/${id}/menu`)
-            .map(res => res.json())
-            .catch(ErrorHandler.handleError);
+        return this.http.get<MenuItem[]>(`${MEATT_API}/restaurants/${id}/menu`)
     }
 }
